+++ pypgstac/pypgstac/migrations/pgstac.0.6.6-0.6.7.sql
@@ -1,140 +1,185 @@
 SET SEARCH_PATH to pgstac, public;
 set check_function_bodies = off;
 
-
-INSERT INTO queryables (name, definition) VALUES
-('id', '{"title": "Item ID","description": "Item identifier","$ref": "https://schemas.stacspec.org/v1.0.0/item-spec/json-schema/item.json#/definitions/core/allOf/2/properties/id"}'),
-('datetime','{"description": "Datetime","type": "string","title": "Acquired","format": "date-time","pattern": "(\\+00:00|Z)$"}'),
-('geometry', '{"title": "Item Geometry","description": "Item Geometry","$ref": "https://geojson.org/schema/Feature.json"}')
-ON CONFLICT DO NOTHING;
-
-INSERT INTO queryables (name, definition, property_wrapper, property_index_type) VALUES
-('eo:cloud_cover','{"$ref": "https://stac-extensions.github.io/eo/v1.0.0/schema.json#/definitions/fieldsproperties/eo:cloud_cover"}','to_int','BTREE')
-ON CONFLICT DO NOTHING;
-
-CREATE OR REPLACE FUNCTION pgstac.get_queryables(_collection text DEFAULT NULL::text)
- RETURNS jsonb
- LANGUAGE sql
-AS $function$
-    SELECT
-        CASE
-            WHEN _collection IS NULL THEN get_queryables(NULL::text[])
-            ELSE get_queryables(ARRAY[_collection])
-        END
-    ;
-$function$
-;
-
-CREATE OR REPLACE FUNCTION pgstac.get_queryables(_collection_ids text[] DEFAULT NULL::text[])
- RETURNS jsonb
- LANGUAGE sql
- STABLE
-AS $function$
-    SELECT
-        jsonb_build_object(
-            '$schema', 'http://json-schema.org/draft-07/schema#',
-            '$id', 'https://example.org/queryables',
-            'type', 'object',
-            'title', 'Stac Queryables.',
-            'properties', jsonb_object_agg(
-                name,
-                definition
-            )
-        )
-        FROM queryables
-        WHERE
-            _collection_ids IS NULL OR
-            cardinality(_collection_ids) = 0 OR
-            collection_ids IS NULL OR
-            _collection_ids && collection_ids
-        ;
-$function$
-;
-
-CREATE OR REPLACE FUNCTION pgstac.missing_queryables(_collection text, _tablesample integer DEFAULT 5)
- RETURNS TABLE(collection text, name text, definition jsonb, property_wrapper text)
- LANGUAGE plpgsql
-AS $function$
+CREATE OR REPLACE FUNCTION search(_search jsonb = '{}'::jsonb) RETURNS jsonb AS $$
 DECLARE
-    q text;
-    _partition text;
-    explain_json json;
-    psize bigint;
+    searches searches%ROWTYPE;
+    _where text;
+    token_where text;
+    full_where text;
+    orderby text;
+    query text;
+    token_type text := substr(_search->>'token',1,4);
+    _limit int := coalesce((_search->>'limit')::int, 10);
+    curs refcursor;
+    cntr int := 0;
+    iter_record items%ROWTYPE;
+    first_record jsonb;
+    first_item items%ROWTYPE;
+    last_item items%ROWTYPE;
+    last_record jsonb;
+    out_records jsonb := '[]'::jsonb;
+    prev_query text;
+    next text;
+    prev_id text;
+    has_next boolean := false;
+    has_prev boolean := false;
+    prev text;
+    total_count bigint;
+    context jsonb;
+    collection jsonb;
+    includes text[];
+    excludes text[];
+    exit_flag boolean := FALSE;
+    batches int := 0;
+    timer timestamptz := clock_timestamp();
+    pstart timestamptz;
+    pend timestamptz;
+    pcurs refcursor;
+    search_where search_wheres%ROWTYPE;
+    id text;
 BEGIN
-    SELECT format('_items_%s', key) INTO _partition FROM collections WHERE id=_collection;
-
-    EXECUTE format('EXPLAIN (format json) SELECT 1 FROM %I;', _partition)
-    INTO explain_json;
-    psize := explain_json->0->'Plan'->'Plan Rows';
-    IF _tablesample * .01 * psize < 10 THEN
-        _tablesample := 100;
-    END IF;
-    RAISE NOTICE 'Using tablesample % to find missing queryables from % % that has ~% rows', _tablesample, _collection, _partition, psize;
-
-    q := format(
-        $q$
-            WITH q AS (
-                SELECT * FROM queryables
-                WHERE
-                    collection_ids IS NULL
-                    OR %L = ANY(collection_ids)
-            ), t AS (
-                SELECT
-                    content->'properties' AS properties
-                FROM
-                    %I
-                TABLESAMPLE SYSTEM(%L)
-            ), p AS (
-                SELECT DISTINCT ON (key)
-                    key,
-                    value
-                FROM t
-                JOIN LATERAL jsonb_each(properties) ON TRUE
-                LEFT JOIN q ON (q.name=key)
-                WHERE q.definition IS NULL
-            )
-            SELECT
-                %L,
-                key,
-                jsonb_build_object('type',jsonb_typeof(value)) as definition,
-                CASE jsonb_typeof(value)
-                    WHEN 'number' THEN 'to_float'
-                    WHEN 'array' THEN 'to_text_array'
-                    ELSE 'to_text'
-                END
-            FROM p;
-        $q$,
-        _collection,
-        _partition,
-        _tablesample,
-        _collection
-    );
-    RETURN QUERY EXECUTE q;
-END;
-$function$
-;
-
-CREATE OR REPLACE FUNCTION pgstac.missing_queryables(_tablesample integer DEFAULT 5)
- RETURNS TABLE(collection_ids text[], name text, definition jsonb, property_wrapper text)
- LANGUAGE sql
-AS $function$
+CREATE TEMP TABLE results (content jsonb) ON COMMIT DROP;
+-- if ids is set, short circuit and just use direct ids query for each id
+-- skip any paging or caching
+-- hard codes ordering in the same order as the array of ids
+IF _search ? 'ids' THEN
+    INSERT INTO results
     SELECT
-        array_agg(collection),
-        name,
-        definition,
-        property_wrapper
-    FROM
-        collections
-        JOIN LATERAL
-        missing_queryables(id, _tablesample) c
-        ON TRUE
-    GROUP BY
-        2,3,4
-    ORDER BY 2,1
+        CASE WHEN _search->'conf'->>'nohydrate' IS NOT NULL AND (_search->'conf'->>'nohydrate')::boolean = true THEN
+            content_nonhydrated(items, _search->'fields')
+        ELSE
+            content_hydrate(items, _search->'fields')
+        END
+    FROM items WHERE
+        items.id = ANY(to_text_array(_search->'ids'))
+        AND
+            CASE WHEN _search ? 'collections' THEN
+                items.collection = ANY(to_text_array(_search->'collections'))
+            ELSE TRUE
+            END
+    ORDER BY items.datetime desc, items.id desc
     ;
-$function$
-;
+    SELECT INTO total_count count(*) FROM results;
+ELSE
+    searches := search_query(_search);
+    _where := searches._where;
+    orderby := searches.orderby;
+    search_where := where_stats(_where);
+    total_count := coalesce(search_where.total_count, search_where.estimated_count);
 
+    IF token_type='prev' THEN
+        token_where := get_token_filter(_search, null::jsonb);
+        orderby := sort_sqlorderby(_search, TRUE);
+    END IF;
+    IF token_type='next' THEN
+        token_where := get_token_filter(_search, null::jsonb);
+    END IF;
+
+    full_where := concat_ws(' AND ', _where, token_where);
+    RAISE NOTICE 'FULL QUERY % %', full_where, clock_timestamp()-timer;
+    timer := clock_timestamp();
+
+    FOR query IN SELECT partition_queries(full_where, orderby, search_where.partitions) LOOP
+        timer := clock_timestamp();
+        query := format('%s LIMIT %s', query, _limit + 1);
+        RAISE NOTICE 'Partition Query: %', query;
+        batches := batches + 1;
+        -- curs = create_cursor(query);
+        OPEN curs FOR EXECUTE query;
+        LOOP
+            FETCH curs into iter_record;
+            EXIT WHEN NOT FOUND;
+            cntr := cntr + 1;
+
+            IF _search->'conf'->>'nohydrate' IS NOT NULL AND (_search->'conf'->>'nohydrate')::boolean = true THEN
+                last_record := content_nonhydrated(iter_record, _search->'fields');
+            ELSE
+                last_record := content_hydrate(iter_record, _search->'fields');
+            END IF;
+            last_item := iter_record;
+            IF cntr = 1 THEN
+                first_item := last_item;
+                first_record := last_record;
+            END IF;
+            IF cntr <= _limit THEN
+                INSERT INTO results (content) VALUES (last_record);
+            ELSIF cntr > _limit THEN
+                has_next := true;
+                exit_flag := true;
+                EXIT;
+            END IF;
+        END LOOP;
+        CLOSE curs;
+        RAISE NOTICE 'Query took %.', clock_timestamp()-timer;
+        timer := clock_timestamp();
+        EXIT WHEN exit_flag;
+    END LOOP;
+    RAISE NOTICE 'Scanned through % partitions.', batches;
+END IF;
+
+SELECT jsonb_agg(content) INTO out_records FROM results WHERE content is not NULL;
+
+DROP TABLE results;
+
+
+-- Flip things around if this was the result of a prev token query
+IF token_type='prev' THEN
+    out_records := flip_jsonb_array(out_records);
+    first_item := last_item;
+    first_record := last_record;
+END IF;
+
+-- If this query has a token, see if there is data before the first record
+IF _search ? 'token' THEN
+    prev_query := format(
+        'SELECT 1 FROM items WHERE %s LIMIT 1',
+        concat_ws(
+            ' AND ',
+            _where,
+            trim(get_token_filter(_search, to_jsonb(first_item)))
+        )
+    );
+    RAISE NOTICE 'Query to get previous record: % --- %', prev_query, first_record;
+    EXECUTE prev_query INTO has_prev;
+    IF FOUND and has_prev IS NOT NULL THEN
+        RAISE NOTICE 'Query results from prev query: %', has_prev;
+        has_prev := TRUE;
+    END IF;
+END IF;
+has_prev := COALESCE(has_prev, FALSE);
+
+IF has_prev THEN
+    prev := out_records->0->>'id';
+END IF;
+IF has_next OR token_type='prev' THEN
+    next := out_records->-1->>'id';
+END IF;
+
+IF context(_search->'conf') != 'off' THEN
+    context := jsonb_strip_nulls(jsonb_build_object(
+        'limit', _limit,
+        'matched', total_count,
+        'returned', coalesce(jsonb_array_length(out_records), 0)
+    ));
+ELSE
+    context := jsonb_strip_nulls(jsonb_build_object(
+        'limit', _limit,
+        'returned', coalesce(jsonb_array_length(out_records), 0)
+    ));
+END IF;
+
+collection := jsonb_build_object(
+    'type', 'FeatureCollection',
+    'features', coalesce(out_records, '[]'::jsonb),
+    'next', next,
+    'prev', prev,
+    'context', context
+);
+
+RETURN collection;
+END;
+$$ LANGUAGE PLPGSQL SECURITY DEFINER SET SEARCH_PATH TO pgstac, public;
 
 
 SELECT set_version('0.6.7');